api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.6.1